Store uuid in a variable to be used in note object being pushed to the notes array...
Redirect to edit.html page when 'Create Note' button is clicked. Removed renderNotes function call as there was no need to render to the DOM
as the page was being redirected. UUID of new note being created  was also injected into the redirect url to aid in editing the correct note.